readBmp280/requirements.txt
paho-mqtt
smbus
bme280